Move JSPM files inside www/
All the JSPM packages have to be available at runtime. config.js needs
to be available both to the tooling and at runtime. The tooling needs to
see package.json in the same folder as config.js: Thus we have two
package.json, one with the jspm deps inside www/ and one in mobile/ with
npm info. When I write a version bump gulp task, I'll have it update
both.