Implement complete back button logic
The backbutton event handler is only attached when the user is not on a
view marked "home". This is because, in order for the back button to
exit the app from the main screen in a platform-appropriate way, the
handler has to be unattached.

When the handler does fire, it first attempts to call
router.navigateBack(). If there is no navigation from this action (read,
the app's been resumed from tombstoning), it instead navigates to the
route's defaultBack property (set in the config object).