Hook up askQuestion
Hooked up askQuestion to the button in courseHome.

In order to deal with a quirk of the framework, I had to split the
ask/edit component into two routes (they still point to the same VM).

Edited questionBox to bind the title, rather than hard-coding it.
Thought it would be necessary. Might not have been in the end but I left
it like that.