Preserve state of question boxes across navigations
In order to make it possible for the question boxes to still be zoomed
when you navigate back from a question page, I moved the zoom toggling
out of the element's view-model and into the courseHome view-model,
(making `showList` into a bindable). This way, we can intelligently
remember state, listening for the activate and deactivate hooks to
manage it.