Go back to Course Home and delete question after answering
After the user submits an answer to a quiz question, delete the question
and return to course home.

NOTE: There is a BIZARRE bug regarding Aurelia, Cordova, and Windows
Phone leading to the following fact: If you delete or replace a Map
element, and that Map has EVER been iterated over in a repeat.for
without that repeat.for being removed from the DOM, then the app will
crash.

SOLUTION: Before either leaving a screen that has a repeat.for over a
map, or deleting an element from a map, use an if.bind on a wrapper div
to remove that repeat.for from the DOM. You can show the repeat again
immediately after you make the change.