Sign in/out
Forgot to commit this the other day. Functional signing in and out,
complete with fix for broken redirect.

If for some reason you get a PERMISSION DENIED error when trying to make
an account, just restart the app. Haven't figured out what's going on
there yet... You will also need to delete the half-created user from the
server (it will be registered, but have no profile object).