Fix issue with reconstructing EnhancedReactiveCollections
Firebase maintains an internal cache. If you register a callback and
something relevant is in the cache, it calls your callback right away
and *synchronously*. This breaks inheriting from ReactiveCollection
because super() has to come before your setup code.

Fix by wrapping the contents of _listenToQuery() in a setTimeout-0.

This fixes the issue with Add Course not loading right the second time.