Fixes for various CSS bugs
Tired me used somewhat hacky methods to get the CSS for the
QuestionBoxes looking right. I did the fiddling in Firefox OS and it
looks like WP has some bugs with what I did. This should fix those. One
of the bugs was the giant titlebar so that's probably fixed on Android
now, too.